move Any, AnyRef to bokeh.core.property.any
